Conflicts:
	library/src/com/nostra13/universalimageloader/core/ImageLoadingInfo.java
	library/src/com/nostra13/universalimageloader/core/LoadAndDisplayImageTask.java

+++ library/src/com/nostra13/universalimageloader/cache/disc/LimitedDiscCache.java
@@ -29,13 +29,15 @@ import java.util.concurrent.atomic.AtomicInteger;
 /**
  * Abstract disc cache limited by some parameter. If cache exceeds specified limit then file with the most oldest last
  * usage date will be deleted.
- * 
+ *
  * @author Sergey Tarasevich (nostra13[at]gmail[dot]com)
- * @since 1.0.0
  * @see BaseDiscCache
  * @see FileNameGenerator
+ * @since 1.0.0
  */
 public abstract class LimitedDiscCache extends BaseDiscCache {
+
+	private static final int INVALID_SIZE = -1;
 
 	private final AtomicInteger cacheSize;
 
@@ -44,21 +46,21 @@ public abstract class LimitedDiscCache extends BaseDiscCache {
 	private final Map<File, Long> lastUsageDates = Collections.synchronizedMap(new HashMap<File, Long>());
 
 	/**
-	 * @param cacheDir Directory for file caching. <b>Important:</b> Specify separate folder for cached files. It's
-	 *            needed for right cache limit work.
+	 * @param cacheDir  Directory for file caching. <b>Important:</b> Specify separate folder for cached files. It's
+	 *                  needed for right cache limit work.
 	 * @param sizeLimit Cache limit value. If cache exceeds this limit then file with the most oldest last usage date
-	 *            will be deleted.
+	 *                  will be deleted.
 	 */
 	public LimitedDiscCache(File cacheDir, int sizeLimit) {
 		this(cacheDir, DefaultConfigurationFactory.createFileNameGenerator(), sizeLimit);
 	}
 
 	/**
-	 * @param cacheDir Directory for file caching. <b>Important:</b> Specify separate folder for cached files. It's
-	 *            needed for right cache limit work.
+	 * @param cacheDir          Directory for file caching. <b>Important:</b> Specify separate folder for cached files. It's
+	 *                          needed for right cache limit work.
 	 * @param fileNameGenerator Name generator for cached files
-	 * @param sizeLimit Cache limit value. If cache exceeds this limit then file with the most oldest last usage date
-	 *            will be deleted.
+	 * @param sizeLimit         Cache limit value. If cache exceeds this limit then file with the most oldest last usage date
+	 *                          will be deleted.
 	 */
 	public LimitedDiscCache(File cacheDir, FileNameGenerator fileNameGenerator, int sizeLimit) {
 		super(cacheDir, fileNameGenerator);
@@ -90,9 +92,10 @@ public abstract class LimitedDiscCache extends BaseDiscCache {
 	public void put(String key, File file) {
 		int valueSize = getSize(file);
 		int curCacheSize = cacheSize.get();
+
 		while (curCacheSize + valueSize > sizeLimit) {
 			int freedSize = removeNext();
-			if (freedSize == 0) break; // cache is empty (have nothing to delete)
+			if (freedSize == INVALID_SIZE) break; // cache is empty (have nothing to delete)
 			curCacheSize = cacheSize.addAndGet(-freedSize);
 		}
 		cacheSize.addAndGet(valueSize);
@@ -123,9 +126,8 @@ public abstract class LimitedDiscCache extends BaseDiscCache {
 	/** Remove next file and returns it's size */
 	private int removeNext() {
 		if (lastUsageDates.isEmpty()) {
-			return 0;
+			return INVALID_SIZE;
 		}
-
 		Long oldestUsage = null;
 		File mostLongUsedFile = null;
 		Set<Entry<File, Long>> entries = lastUsageDates.entrySet();
@@ -144,9 +146,16 @@ public abstract class LimitedDiscCache extends BaseDiscCache {
 			}
 		}
 
-		int fileSize = getSize(mostLongUsedFile);
-		if (mostLongUsedFile.delete()) {
-			lastUsageDates.remove(mostLongUsedFile);
+		int fileSize = 0;
+		if (mostLongUsedFile != null) {
+			if (mostLongUsedFile.exists()) {
+				fileSize = getSize(mostLongUsedFile);
+				if (mostLongUsedFile.delete()) {
+					lastUsageDates.remove(mostLongUsedFile);
+				}
+			} else {
+				lastUsageDates.remove(mostLongUsedFile);
+			}
 		}
 		return fileSize;
 	}
